+++ packages/core/badge/src/index.js
@@ -4,5 +4,4 @@ export { default } from './components';
 export { Container } from './components/Container';
 export { Format } from './components/Format';
 export { backgroundColors, textColors, theme } from './theme';
-
 export type { ThemeAppearance, ThemeProps } from './theme';