+++ packages/core/inline-edit/examples/02-select-example.js
@@ -64,7 +64,7 @@ export default class SelectExample extends Component<void, State> {
 
   render() {
     return (
-      <div>
+      <div style={{ padding: '0 16px' }}>
         <InlineEditor
           label="With Multi Select Edit View"
           disableEditViewFieldBase