+++ packages/core/lozenge/src/Lozenge/styledContainer.js
@@ -9,13 +9,12 @@ export default styled.span`
   `};
   border-radius: ${akBorderRadius};
   box-sizing: border-box;
-
   display: inline-block;
   font-size: 11px;
   font-weight: 700;
   line-height: 1;
+  max-width: 100%;
   padding: 2px 0 3px 0;
   text-transform: uppercase;
   vertical-align: baseline;
-  max-width: 100%;
 `;