+++ packages/core/navigation-next/src/components/Item/styles.js
@@ -29,7 +29,7 @@ const baseStyles = {
     display: 'flex',
     flexShrink: 0,
     fontSize: 'inherit',
-    height: `${gridSize * 5}`,
+    height: gridSize * 5,
     outline: 'none',
     textAlign: 'left',
     textDecoration: 'none',
@@ -74,34 +74,34 @@ const baseStyles = {
 const layoutStyles = {
   compact: {
     itemBase: {
-      paddingRight: `${gridSize}px`,
-      paddingLeft: `${gridSize}px`,
+      paddingRight: gridSize,
+      paddingLeft: gridSize,
     },
     beforeWrapper: {
-      marginRight: `${gridSize}`,
+      marginRight: gridSize,
     },
     subTextWrapper: {
-      fontSize: `${10}px`,
+      fontSize: '10px',
       lineHeight: 1.2,
     },
     afterWrapper: {
-      marginLeft: `${gridSize}`,
+      marginLeft: gridSize,
     },
   },
   default: {
     itemBase: {
-      paddingRight: `${gridSize * 1.5}px`,
-      paddingLeft: `${gridSize * 1.5}px`,
+      paddingLeft: gridSize * 1.5,
+      paddingRight: gridSize * 1.5,
     },
     beforeWrapper: {
-      marginRight: `${gridSize * 2}`,
+      marginRight: gridSize * 2,
     },
     subTextWrapper: {
-      fontSize: `${12}px`,
+      fontSize: '12px',
       lineHeight: 14 / 12,
     },
     afterWrapper: {
-      marginLeft: `${gridSize * 2}`,
+      marginLeft: gridSize * 2,
     },
   },
 };
@@ -122,75 +122,82 @@ export default ({ product }: ModeColors) => ({
   isHover,
   isSelected,
   spacing,
-}: ItemPresentationProps) => ({
-  container: {
-    itemBase: {
-      ...baseStyles.itemBase,
-      ...layoutStyles[spacing].itemBase,
-      backgroundColor: getItemBackgroundColor(
-        {
-          default: colors.N20,
-          hint: colors.N30,
-          interact: colors.B50,
-          static: colors.N30,
-        },
-        {
-          isActive,
-          isHover,
-          isSelected,
-        },
-      ),
+}: ItemPresentationProps) => {
+  const containerTextColor = isActive || isSelected ? colors.B400 : colors.N500;
+  const containerBackgroundColor = getItemBackgroundColor(
+    {
+      default: colors.N20,
+      hint: colors.N30,
+      interact: colors.B50,
+      static: colors.N30,
     },
-    beforeWrapper: {
-      ...baseStyles.beforeWrapper,
-      ...layoutStyles[spacing].beforeWrapper,
-      color: isSelected ? colors.B400 : colors.N500,
+    {
+      isActive,
+      isHover,
+      isSelected,
     },
-    contentWrapper: baseStyles.contentWrapper,
-    textWrapper: {
-      ...baseStyles.textWrapper,
-      color: isSelected ? colors.B400 : colors.N500,
+  );
+  const productBackgroundColor = getItemBackgroundColor(product.background, {
+    isActive,
+    isHover,
+    isSelected,
+  });
+  return {
+    container: {
+      itemBase: {
+        ...baseStyles.itemBase,
+        ...layoutStyles[spacing].itemBase,
+        backgroundColor: containerBackgroundColor,
+        fill: containerBackgroundColor,
+      },
+      beforeWrapper: {
+        ...baseStyles.beforeWrapper,
+        ...layoutStyles[spacing].beforeWrapper,
+        color: containerTextColor,
+      },
+      contentWrapper: baseStyles.contentWrapper,
+      textWrapper: {
+        ...baseStyles.textWrapper,
+        color: containerTextColor,
+      },
+      subTextWrapper: {
+        ...baseStyles.subTextWrapper,
+        ...layoutStyles[spacing].subTextWrapper,
+        color: colors.N200,
+      },
+      afterWrapper: {
+        ...baseStyles.afterWrapper,
+        ...layoutStyles[spacing].afterWrapper,
+        color: colors.N500,
+      },
     },
-    subTextWrapper: {
-      ...baseStyles.subTextWrapper,
-      ...layoutStyles[spacing].subTextWrapper,
-      color: colors.N200,
+    product: {
+      itemBase: {
+        ...baseStyles.itemBase,
+        ...layoutStyles[spacing].itemBase,
+        backgroundColor: productBackgroundColor,
+        fill: productBackgroundColor,
+      },
+      beforeWrapper: {
+        ...baseStyles.beforeWrapper,
+        ...layoutStyles[spacing].beforeWrapper,
+        color: product.text.default,
+      },
+      contentWrapper: baseStyles.contentWrapper,
+      textWrapper: {
+        ...baseStyles.textWrapper,
+        color: product.text.default,
+      },
+      subTextWrapper: {
+        ...baseStyles.subTextWrapper,
+        ...layoutStyles[spacing].subTextWrapper,
+        color: product.text.subtle,
+      },
+      afterWrapper: {
+        ...baseStyles.afterWrapper,
+        ...layoutStyles[spacing].afterWrapper,
+        color: product.text.default,
+      },
     },
-    afterWrapper: {
-      ...baseStyles.afterWrapper,
-      ...layoutStyles[spacing].afterWrapper,
-      color: colors.N500,
-    },
-  },
-  product: {
-    itemBase: {
-      ...baseStyles.itemBase,
-      ...layoutStyles[spacing].itemBase,
-      backgroundColor: getItemBackgroundColor(product.background, {
-        isActive,
-        isHover,
-        isSelected,
-      }),
-    },
-    beforeWrapper: {
-      ...baseStyles.beforeWrapper,
-      ...layoutStyles[spacing].beforeWrapper,
-      color: product.text.default,
-    },
-    contentWrapper: baseStyles.contentWrapper,
-    textWrapper: {
-      ...baseStyles.textWrapper,
-      color: product.text.default,
-    },
-    subTextWrapper: {
-      ...baseStyles.subTextWrapper,
-      ...layoutStyles[spacing].subTextWrapper,
-      color: product.text.subtle,
-    },
-    afterWrapper: {
-      ...baseStyles.afterWrapper,
-      ...layoutStyles[spacing].afterWrapper,
-      color: product.text.default,
-    },
-  },
-});
+  };
+};