+++ packages/editor/adf-utils/src/builders/nodes/paragraph.ts
@@ -1,25 +1,4 @@
-import {
-  ActionDefinition,
-  CodeDefinition,
-  DateDefinition,
-  EmDefinition,
-  EmojiDefinition,
-  HardBreakDefinition,
-  InlineExtensionDefinition,
-  LinkDefinition,
-  MarksObject,
-  MentionDefinition,
-  PlaceholderDefinition,
-  StrikeDefinition,
-  StrongDefinition,
-  SubSupDefinition,
-  TextColorDefinition,
-  TextDefinition,
-  UnderlineDefinition,
-  ParagraphDefinition,
-  Inline,
-  InlineCardDefinition,
-} from '@atlaskit/editor-common';
+import { ParagraphDefinition, Inline } from '@atlaskit/editor-common';
 import { createTextNodes } from '../utils/create-text-nodes';
 
 export const paragraph = (
@@ -28,26 +7,3 @@ export const paragraph = (
   type: 'paragraph',
   content: createTextNodes(content),
 });
-
-export {
-  ActionDefinition,
-  CodeDefinition,
-  DateDefinition,
-  EmDefinition,
-  EmojiDefinition,
-  HardBreakDefinition,
-  InlineExtensionDefinition,
-  LinkDefinition,
-  MarksObject,
-  MentionDefinition,
-  PlaceholderDefinition,
-  StrikeDefinition,
-  StrongDefinition,
-  SubSupDefinition,
-  TextColorDefinition,
-  TextDefinition,
-  UnderlineDefinition,
-  ParagraphDefinition,
-  Inline,
-  InlineCardDefinition,
-};