+++ packages/editor/editor-wikimarkup-transformer/src/parser/tokenize/index.ts
@@ -1,5 +1,6 @@
 import { Node as PMNode, Schema } from 'prosemirror-model';
-import { attachment } from './attachment';
+import { media } from './media';
+import { fileLink } from './file-link';
 import { blockquote } from './blockquote';
 import { citation } from './citation';
 import { deleted } from './deleted';
@@ -33,7 +34,8 @@ export enum TokenType {
   STRING = 'STRING',
   LINK_FORMAT = 'LINK_FORMAT',
   LINK_TEXT = 'LINK_TEXT',
-  ATTACHMENT = 'ATTACHMENT',
+  MEDIA = 'MEDIA',
+  FLIE_LINK = 'FILE_LINK',
   HEADING = 'HEADING',
   LIST = 'LIST',
   TABLE = 'TABLE',
@@ -88,7 +90,8 @@ const tokenToTokenParserMapping: {
   [TokenType.LINK_FORMAT]: linkFormat,
   [TokenType.LINK_TEXT]: linkText,
   [TokenType.HEADING]: heading,
-  [TokenType.ATTACHMENT]: attachment,
+  [TokenType.MEDIA]: media,
+  [TokenType.FLIE_LINK]: fileLink,
   [TokenType.LIST]: list,
   [TokenType.QUOTE]: blockquote,
   [TokenType.TABLE]: table,