+++ packages/core/lozenge/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/lozenge
 
+## 6.2.0
+- [minor] Use new theme API. Adds "theme" prop and exports new theme types. [4b36fd6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4b36fd6)
+
 ## 6.1.8
 - [patch] AK-5321 Lozenge with maxWidth should be constrained by container width [969233e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/969233e)