+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/global-search
 
+## 5.5.11
+- [patch] Update TS to 3.0 [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+
 ## 5.5.10
 - [patch] Updated dependencies [d5a043a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d5a043a)
   - @atlaskit/icon@13.8.1